re-implemented API's to account for unauthorized access

+++ src/main/java/com/microservice/gpstracker/Vehicle/VehicleController.java
@@ -1,11 +1,15 @@
 package com.microservice.gpstracker.Vehicle;
 
 import com.microservice.gpstracker.Owner.OwnerService;
+import com.microservice.gpstracker.util.AuthorizeUtility;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 
+@SuppressWarnings("OptionalGetWithoutIsPresent")
 @RestController
 @RequestMapping("/vehicle/{owner_id}")
 public class VehicleController {
@@ -16,14 +20,21 @@ public class VehicleController {
     @Autowired
     private OwnerService ownerService;
 
+    @Autowired
+    private AuthorizeUtility utility;
+
     @GetMapping("/all-vehicles")
     public List<VehicleModel> getAllVehiclesByOwnerID(@PathVariable("owner_id") Long owner_id) {
         return this.vehicleService.findAllVehicleByOwnerModelOwnerId(owner_id);
     }
 
-    @GetMapping("/get-vehicle/{vehicle_id}")
-    public VehicleModel getVehicleById(@PathVariable("vehicle_id") Long vehicle_id) {
-        return this.vehicleService.getVehicleById(vehicle_id).get();
+    @GetMapping("/get/{vehicle_id}")
+    public ResponseEntity<Object> getVehicleById(@PathVariable("vehicle_id") Long owner_id,
+                                                 @PathVariable("vehicle_id") Long vehicle_id) {
+        VehicleModel vehicle = this.vehicleService.getVehicleById(vehicle_id).get();
+        if (utility.checkIfValid(owner_id, vehicle))
+            return new ResponseEntity<>(vehicle, HttpStatus.OK);
+        return new ResponseEntity<>("Unauthorised", HttpStatus.UNAUTHORIZED);
     }
 
     @PostMapping("/save")
@@ -39,9 +50,16 @@ public class VehicleController {
         return this.vehicleService.deleteAllVehicles();
     }
 
-    @DeleteMapping("/delete/{id}")
-    public String deleteVehicleById(@PathVariable("id") Long vehicle_id) {
-        return this.vehicleService.deleteVehicleById(vehicle_id);
+    @DeleteMapping("/delete/{vehicle_id}")
+    public ResponseEntity<Object> deleteVehicleById(@PathVariable("owner_id") Long owner_id,
+                                                    @PathVariable("vehicle_id") Long vehicle_id) {
+        VehicleModel vehicle = this.vehicleService.getVehicleById(vehicle_id).get();
+        if (utility.checkIfValid(owner_id, vehicle)) {
+            this.vehicleService.deleteVehicleById(vehicle_id);
+            return new ResponseEntity<>("Deleted \n" + vehicle, HttpStatus.OK);
+        }
+        return new ResponseEntity<>("Unauthorised", HttpStatus.UNAUTHORIZED);
+
     }
 
 }
